Move addresses and hook into rct2 sub directory
