+++ openrct2.vcxproj
@@ -88,7 +88,7 @@
     <None Include="openrct2.exe" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="src\addresses.c" />
+    <ClCompile Include="src\rct2\addresses.c" />
     <ClCompile Include="src\audio\audio.c" />
     <ClCompile Include="src\audio\mixer.cpp" />
     <ClCompile Include="src\cheats.c" />
@@ -140,7 +140,7 @@
     <ClCompile Include="src\drawing\string.c" />
     <ClCompile Include="src\editor.c" />
     <ClCompile Include="src\game.c" />
-    <ClCompile Include="src\hook.c" />
+    <ClCompile Include="src\rct2\hook.c" />
     <ClCompile Include="src\image_io.c" />
     <ClCompile Include="src\input.c" />
     <ClCompile Include="src\interface\chat.c" />
@@ -422,7 +422,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resources\resource.h" />
-    <ClInclude Include="src\addresses.h" />
+    <ClInclude Include="src\rct2\addresses.h" />
     <ClInclude Include="src\audio\audio.h" />
     <ClInclude Include="src\audio\mixer.h" />
     <ClInclude Include="src\cheats.h" />
@@ -476,7 +476,7 @@
     <ClInclude Include="src\drawing\lightfx.h" />
     <ClInclude Include="src\editor.h" />
     <ClInclude Include="src\game.h" />
-    <ClInclude Include="src\hook.h" />
+    <ClInclude Include="src\rct2\hook.h" />
     <ClInclude Include="src\image_io.h" />
     <ClInclude Include="src\input.h" />
     <ClInclude Include="src\interface\chat.h" />
