+++ src/rct2/addresses.c
